pyproject.toml
[project]
name = "ai-webscraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.13",
    "langchain-anthropic>=0.3.15",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
    "streamlit>=1.46.0",
    "uvicorn>=0.34.3",
]
